requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
PyJWT[crypto]==2.8.0
python-dotenv==1.0.1
cryptography
pytest==7.4.0
pytest-asyncio==0.22.0
python-jose[cryptography]==3.1.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
sqlalchemy
pyodbc